requirements.txt
# API
fastapi
uvicorn
python-multipart
pydantic>=2.7
python-dotenv

# LLMs
google-generativeai>=0.7.2
openai>=1.35.0

# PDF
pypdf
pdfminer.six
# Optional: advanced extractor (not required by this PoC)
# PyMuPDF

# DOCX
python-docx

# NER / NLP (used by your chat/docx parsers)
transformers
torch
spacy
pandas
numpy
scikit-learn
regex
python-dateutil
